moved the modal code to its own component
